adv req 1-5

+++ Assets/Scripts/Utils/IdScriptableObject.cs
@@ -2,7 +2,7 @@
 
 namespace ThisGame.Utils {
     public class IdScriptableObject : ScriptableObject {
-        public ushort id;
+        public uint id;
         public string nameid;
         public new string name;
         [TextArea] public string desc;
